(#15) Allow large raw file uploads
Previously only small files were able to be uploaded to a Raw repository in
Nexus using this module.

With this commit several private helper functions were added to smartly determine
the size of the file being uploaded, and dynamically chunking the file and sending
individual requests to the API endpoint. This results in the ability for a file of
any size to successfully be uploaded to your Raw repository.