(#23) Fix pagination bug in Get-NexusComponent
Prior to this commit a max of 20 results was returned due to a bug in how
continuation tokens were handled.

With this commit continuation tokens are properly handled, and results are
stored until no continuation token is provided by the response, and then the
results are returned to the user.

Also with this commit the results are returned as proper objects to allow pipeline
operations to work as expected. Previously things like Select-Object returned null results.